Application and model name changes to start Orchard application
Application was a "fork" (import) of previous webapp for a movie database.
The name of the application and models needed to start to change to clearly see the
intended outcomes and use of each. This was a search and replace for application and model.
There are also some changes to the main model for the database.